PLC-39: data mapping PLCnext 2 OpenEMS

+++ io.openems.edge.phoenixcontact.plcnext/src/io/openems/edge/phoenixcontact/plcnext/PlcNextDevice.java
@@ -10,22 +10,22 @@ import io.openems.edge.meter.api.ElectricityMeter;
 public interface PlcNextDevice extends ElectricityMeter, OpenemsComponent {
 
 	public enum ChannelId implements io.openems.edge.common.channel.ChannelId {
-		LINE_VOLTAGE(Doc.of(OpenemsType.FLOAT)//
+		LINE_VOLTAGE(Doc.of(OpenemsType.INTEGER)//
 				.unit(Unit.MILLIVOLT)//
 				.persistencePriority(PersistencePriority.HIGH)),
-		LINE_VOLTAGE_L1(Doc.of(OpenemsType.FLOAT)//
+		LINE_VOLTAGE_L1(Doc.of(OpenemsType.INTEGER)//
 				.unit(Unit.MILLIVOLT)//
 				.persistencePriority(PersistencePriority.HIGH)),
-		LINE_VOLTAGE_L2(Doc.of(OpenemsType.FLOAT)//
+		LINE_VOLTAGE_L2(Doc.of(OpenemsType.INTEGER)//
 				.unit(Unit.MILLIVOLT)//
 				.persistencePriority(PersistencePriority.HIGH)),
-		LINE_VOLTAGE_L3(Doc.of(OpenemsType.FLOAT)//
+		LINE_VOLTAGE_L3(Doc.of(OpenemsType.INTEGER)//
 				.unit(Unit.MILLIVOLT)//
 				.persistencePriority(PersistencePriority.HIGH)),
-		NEUTRAL_CURRENT(Doc.of(OpenemsType.DOUBLE)//
+		NEUTRAL_CURRENT(Doc.of(OpenemsType.INTEGER)//
 				.unit(Unit.MILLIAMPERE)//
 				.persistencePriority(PersistencePriority.HIGH)),
-		APPARENT_POWER(Doc.of(OpenemsType.DOUBLE)//
+		APPARENT_POWER(Doc.of(OpenemsType.INTEGER)//
 				.unit(Unit.WATT)//
 				.persistencePriority(PersistencePriority.HIGH)),
 		POWER_FACTOR(Doc.of(OpenemsType.DOUBLE)//
